manufacturing_Web_Backend/user_data/serializers.py
from rest_framework import serializers
from .models import *


class UserPreferencesSerializer(serializers.ModelSerializer):
    class Meta:
        model = UserPreferences
        fields = '__all__'
