manufacturing_Web_Backend/requirements.txt
django==4.2.7
django-cors-headers==3.6.0
djangorestframework==3.14.0
numpy==1.23.0
scikit-learn==1.2.0
xgboost==2.0.3
pandas==1.4.3
gunicorn==23.0.0
whitenoise==6.7.0